updated files

+++ src/main/java/com/example/control/CourseController.java
@@ -30,7 +30,7 @@ public class CourseController {
     }
 
     @DeleteMapping("delete")
-    public String deleteCourseDetails(@RequestBody int id) {
+    public String deleteCourseDetails(@RequestBody Long id) {
         return courseService.deleteCourse(id);
     }
 }
